third_party/blink/web_tests/platform/mac-mac10.15/external/wpt/css/css-images/image-set/image-set-parsing-expected.txt

This is a testharness.js-based test.
FAIL background-image: image-set(url(example.png) 1x) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set(url('example.png') 1x) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set('example.jpg' 1x) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set(url(example.png) 1x, 'example.png' 2x) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set(url(example.png) 1dppx) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set(url(example.png) 1dpi) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set(url(example.png) 1dpcm, 'example.png' 2x) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set('example.jpeg' 222dpi, url(example.png) 3.5x) [ parsable ] assert_equals: expected true but got false
FAIL background-image: image-set(linear-gradient(black, white) 1x) [ parsable ] assert_equals: expected true but got false
FAIL content: image-set(linear-gradient(black, white) 1x, 'example.png' 4x) [ parsable ] assert_equals: expected true but got false
FAIL content: image-set(url('example.png') 192dpi, linear-gradient(black, white) 1x) [ parsable ] assert_equals: expected true but got false
PASS background-image: image-set(url(example.png) 0x) [ unparsable ]
PASS background-image: image-set(url(example.png) -20x) [ unparsable ]
PASS background-image: image-set('example.jpeg' 92pid url(example.png) 1x) [ unparsable ]
PASS cursor: image-set(linear-gradient(black, white) 1x) [ unparsable ]
Harness: the test ran to completion.
